requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0
wordcloud>=1.9.0
scikit-learn>=1.3.0
gensim>=4.3.0
underthesea>=6.8.0
pyvi>=0.1.1
pyyaml>=6.0
apify-client>=1.7.0
jinja2>=3.1.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
